+++ CHANGELOG.md
@@ -9,7 +9,7 @@ All notable changes to this project will be documented in this file.
 - Start scheduler in out-of-service mode. (#712)
 - Scheduler needs to enter out-of-service mode on first blockchain communication loss.
   This is due to **Nethermind v1.14.7+99775bf7** where filters are lost on restart. (#715)
-- Fix result-proxy URL resolution. (#716)
+- Use Result Proxy URL defined in deal parameters if any, fall back to scheduler default one otherwise. (#716)
 
 ### Dependency Upgrades